ml_algo/requirements.txt
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dotenv-python==0.0.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
Flask==2.2.2
fonttools==4.37.1
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multitasking==0.0.11
mypy-extensions==0.4.3
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.3.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.13.0
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xgboost==1.5.0
yfinance==0.1.74
zipp==3.8.1
